Working but I don't believe the result
The points in polygons function now works, but it returns a df with a lot of duplicates in it, this is not righ. It also warns Left CRS: None, which is not right, because the CRS should be EPSG:27700, which is the same as the right CRS. The answer that's being printed out for the number of people who are not served by public transport in Birmingham is 3.